fix: apply curated originalReleaseYear in the reparse scripts too
scripts/reparse-affected.ts and scripts/reparse-them.ts replace the whole
cache entry for a feed but never called resolveOriginalRelease, so a curated
originalReleaseYear in data/feeds.json was silently discarded the moment
either script ran. Those two are the documented tool for the private feeds —
including Them, the album most likely to need a hand-corrected year — so the
override was broken exactly where it matters most.

Both now resolve the override before writing. The feeds.json lookup that
backfill-album-dates.ts already had moves to scripts/feed-overrides.ts and is
shared by all three; it stays out of lib/album-date.ts because it uses fs and
that module is imported by client components.

Verified end to end through reparse-them.ts, which previously dropped the
value: originalReleaseYear 1999 lands as originalRelease.year 1999, and null
removes the key so the page falls back to the 2026 pubDate year. Backfill is
unchanged and still idempotent — 9 matches, 2 traps rejected, no cache diff.

Also drops a now-redundant `as any` on matchingFeed.originalReleaseYear (Feed
declares the field), hoists AlbumCard's duplicated getDisplayYear call, and
corrects the backfill log line, which counted albums carrying a date rather
than albums changed.
